Refactor multiSelectColumn and move logic per Carson's review
